requirements.txt
web3>=6.0.0
python-telegram-bot>=20.0
requests>=2.28